moved tropomi operator to operators dir
